refactor: Domain 별로 package분리 + DI를 Bean으로 주입하는 것으로 변경
