+++ src/main/java/kr/codesqaud/cafe/article/controller/ArticleController.java
@@ -1,9 +1,8 @@
-package kr.codesqaud.cafe.controller;
+package kr.codesqaud.cafe.article.controller;
 
-import kr.codesqaud.cafe.dto.ArticleFormDto;
-import kr.codesqaud.cafe.domain.Article;
-import kr.codesqaud.cafe.service.ArticleService;
-import org.springframework.beans.factory.annotation.Autowired;
+import kr.codesqaud.cafe.article.domain.Article;
+import kr.codesqaud.cafe.article.dto.ArticleFormDto;
+import kr.codesqaud.cafe.article.service.ArticleService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,8 +13,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 public class ArticleController {
 
     private final ArticleService articleService;
-
-    @Autowired
+    
     public ArticleController(ArticleService articleService) {
         this.articleService = articleService;
     }
