+++ src/main/java/kr/codesqaud/cafe/article/domain/Article.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.domain;
+package kr.codesqaud.cafe.article.domain;
 
 import java.time.LocalDateTime;
