+++ src/main/java/kr/codesqaud/cafe/article/dto/ArticleDetailDto.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.dto;
+package kr.codesqaud.cafe.article.dto;
 
 
 public class ArticleDetailDto {
