+++ src/main/java/kr/codesqaud/cafe/article/repository/ArticleRepository.java
@@ -1,11 +1,11 @@
-package kr.codesqaud.cafe.repository;
+package kr.codesqaud.cafe.article.repository;
 
-import kr.codesqaud.cafe.domain.Article;
+import kr.codesqaud.cafe.article.domain.Article;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
-@Repository
+
 public interface ArticleRepository {
 
     Long save(Article article);
