+++ src/main/java/kr/codesqaud/cafe/user/domain/User.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.domain;
+package kr.codesqaud.cafe.user.domain;
 
 public class User {
     private String userId;
