+++ src/main/java/kr/codesqaud/cafe/user/repository/MemoryUserRepository.java
@@ -1,11 +1,10 @@
-package kr.codesqaud.cafe.repository;
+package kr.codesqaud.cafe.user.repository;
 
-import kr.codesqaud.cafe.domain.User;
-import org.springframework.stereotype.Repository;
+import kr.codesqaud.cafe.user.domain.User;
 
 import java.util.*;
 
-@Repository
+
 public class MemoryUserRepository implements UserRepository {
 
     private static Map<String, User> storage = new HashMap<>();
