+++ src/test/java/kr/codesqaud/cafe/repository/MemoryArticleRepositoryTest.java
@@ -1,6 +1,7 @@
 package kr.codesqaud.cafe.repository;
 
-import kr.codesqaud.cafe.domain.Article;
+import kr.codesqaud.cafe.article.domain.Article;
+import kr.codesqaud.cafe.article.repository.MemoryArticleRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
